backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Async & WebSocket
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Task Queue
celery==5.3.4
redis==5.0.1

# AI/ML & Computer Vision
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
numpy==1.24.3
Pillow==10.2.0
torch==2.1.2
torchvision==0.16.2

# Utilities
python-dotenv==1.0.0
python-decouple==3.8
gunicorn==21.2.0

# Development
ipython==8.20.0
